quaternion: eliminate a couple loads
This macro interface is getting more ugly by getting some of its parameters
from registers and some from memory, but that's the price you pay for
reducing the number of memory accesses.  I've chosen to keep the same
number of macro parameters, even though %2 is no longer used for
consistency and readability.