dec64: add a base64 decoder
Certain bytes may be difficult to transfer over RS-232 to the XT due to
in-band flow control signaling (XON/XOFF), ^Z being EOF on DOS, and the
perennial issue of line ending conversions.

We'll get around this by sending base64 encoded files, but we need a
decoder on the receiving end.  This is a carefully constructed COM program
to decode base64, while not using any of the problematic bytes in its own
code.